Update configuration and refactor tier limits
- Added 'postgres' as a server external package in next.config.ts and updated webpack configuration to handle it.
- Introduced a new file src/lib/stripe/tier-limits.ts to define subscription tier limits, moving the relevant constants and types from src/lib/stripe/usage.ts.
- Updated imports in various components to reference the new tier-limits file.
- Added 'server-only' package to package.json for server-side rendering support.

This refactor improves code organization and clarity regarding subscription tier limits.
