+++ src/lib/stripe/usage.ts
@@ -2,34 +2,9 @@ import { db } from '@/lib/db/drizzle';
 import { learningPlans, usageMetrics, users } from '@/lib/db/schema';
 import { and, eq, sql } from 'drizzle-orm';
 
-/**
- * Subscription tier limits
- */
-export const TIER_LIMITS = {
-  free: {
-    maxActivePlans: 3,
-    monthlyRegenerations: 5,
-    monthlyExports: 10,
-    maxWeeks: 2,
-    maxHours: null,
-  },
-  starter: {
-    maxActivePlans: 10,
-    monthlyRegenerations: 10,
-    monthlyExports: 50,
-    maxWeeks: 8,
-    maxHours: null,
-  },
-  pro: {
-    maxActivePlans: Infinity,
-    monthlyRegenerations: 50,
-    monthlyExports: Infinity,
-    maxWeeks: null, // unlimited
-    maxHours: null,
-  },
-} as const;
+import { TIER_LIMITS, type SubscriptionTier } from './tier-limits';
 
-export type SubscriptionTier = keyof typeof TIER_LIMITS;
+export { TIER_LIMITS, type SubscriptionTier };
 
 // Usage type for incrementing counters
 export type UsageType = 'plan' | 'regeneration' | 'export';
